refactor homolysis reaction
this moves the functions for the homolysis reaction that don't need
`self` into the utils module (which we will have to refactor into more
logical modules at some point).
Further, the reaction module gains a new class `ReactionResult` which,
encompasses both the list of rates as well as the ReactionRecipe,
as we found out that those are usually calculated in one step anyways.

This will need further changes in the RunManager to properly handle
lists of ReactionResults and prepare us for multiple reactions.

refs #23