perf(startup): move plugin discovery out of init!
This considerably speeds up the startup of kimmdy when just asking for
e.g. `kimmdy --help` or when restarting from a checkpoint (in which case
the plugins are already loaded as part of the runmanager, so they don't
have to be discovered again).

The symbols `reaction_plugins` are now imported from `kimmdy.plugins`.
**But** the containsers will be empty until
`kimmdy.plugins.discover_plugins` has been called.
This is only neceassary twice. For `kimmdy --show-plugins` and for
a regular kimmdy run in `cmd.py`.